Add new address format: z-prefixed hex string part 2

+++ include/evmc/evmc.hpp
@@ -304,7 +304,7 @@ constexpr T parse(std::string_view s, std::string_view prefix) noexcept
 }
 
 /// Literal for evmc::address.
-constexpr address operator""_address(const char* s, size_t) noexcept
+constexpr address operator""_address(const char* s, unsigned long) noexcept
 {
     return parse<address>(s, "Z");
 }
